Zynq monitoring generalization and updates for Rev2 (#132)
* zynqmon from YAML
* Update ZynqMonTask.c, retain old version as legacy
* Fix semaphore release issue
* queue issues, hangs in SMBUS
* add a history of last log printouts
* Add break out of infinite loop in I2C state machine setup
* change debug build to minimize size -Os rather than -O0
* reduce code duplication (task stack usage)
* define LOG_USE_COLOR in header, not makefile

+++ common/LocalUart.c
@@ -104,7 +104,7 @@ void UART0Init(uint32_t ui32SysClock)
 
 void UART4Print(const char *str)
 {
-  int size = strlen(str);
+  size_t size = strlen(str);
   for (int i = 0; i < size; ++i) {
     //
     // Write the next character to the UART.
@@ -115,7 +115,7 @@ void UART4Print(const char *str)
 
 void UARTPrint(uint32_t uart_base, const char *str)
 {
-  int size = strlen(str);
+  size_t size = strlen(str);
   for (int i = 0; i < size; ++i) {
     //
     // Write the next character to the UART.
